+++ include/Core/PlaceVisitor.hpp
@@ -8,8 +8,8 @@
 #ifndef PLACEVISITOR_HPP_
 #define PLACEVISITOR_HPP_
 
-#include "Core/QueryParser/Visitor.hpp"
-#include "Core/QueryParser/AST.hpp"
+#include "Core/Query/Visitor.hpp"
+#include "Core/Query/AST.hpp"
 #include <exception>
 #include <vector>
