+++ include/DiscreteVerification/Generators/RangeVisitor.h
@@ -9,7 +9,7 @@
 #define RANGEVISITOR_H
 
 #include "Core/TAPN/TAPN.hpp"
-#include "Core/QueryParser/Visitor.hpp"
+#include "Core/Query/Visitor.hpp"
 
 #include "DiscreteVerification/DataStructures/NonStrictMarkingBase.hpp"
 #include "DiscreteVerification/DataStructures/NonStrictMarking.hpp"
