+++ src/Core/CMakeLists.txt
@@ -1,10 +1,10 @@
 
 
-add_subdirectory(QueryParser)
+add_subdirectory(Query)
 add_library(PlaceVisitor PlaceVisitor.cpp)
-target_link_libraries(PlaceVisitor QueryParser)
+target_link_libraries(PlaceVisitor Query)
 
 add_subdirectory(TAPN)
 
 add_library(Core ArgsParser.cpp VerificationOptions.cpp)
-target_link_libraries(Core QueryParser TAPN)
+target_link_libraries(Core Query TAPN)
