+++ src/Core/Query/CMakeLists.txt
@@ -1,0 +1,3 @@
+
+
+add_library(Query AST.cpp NormalizationVisitor.cpp)
