+++ src/main.cpp
@@ -3,32 +3,79 @@
 #include "Core/TAPN/TAPNModelBuilder.hpp"
 #include "Core/VerificationOptions.hpp"
 #include "Core/ArgsParser.hpp"
-#include "Core/QueryParser/TAPNQueryParser.hpp"
+#include "Core/Query/TAPNQueryParser.hpp"
 #include "Core/TAPN/TimedPlace.hpp"
 #include "DiscreteVerification/DiscreteVerification.hpp"
 #include "DiscreteVerification/DeadlockVisitor.hpp"
+
 #include <unfoldtacpn.h>
 #include <Colored/ColoredPetriNetBuilder.h>
+#include <memory>
 
 using namespace VerifyTAPN;
 using namespace VerifyTAPN::TAPN;
+
+std::unique_ptr<AST::Query> parse_queries(const VerificationOptions& options, const unfoldtacpn::ColoredPetriNetBuilder& builder) {
+    try {
+        auto& queryFile = options.getQueryFile();
+        std::ifstream qfile(queryFile);
+        if (!qfile) {
+            std::cerr << "Could not open " << queryFile << std::endl;
+            std::exit(-1);
+        } else {
+            std::vector<std::pair<unfoldtacpn::PQL::Condition_ptr,std::string>> ast_queries;
+            auto& qnums = options.getQueryNumbers();
+            if (qfile.peek() == '<') { // assumed XML
+                if (qnums.empty()) {
+                    std::cerr << "ERROR: Missing query-indexes for query-file (which is identified as XML-format)" << std::endl;
+                    std::exit(-1);
+                }
+                ast_queries = unfoldtacpn::parse_xml_queries(builder, qfile, qnums);
+            } else {
+                // not xml
+                if (qnums.size() > 0) {
+                    std::cerr << "Querys not provided in XML-format, -q-num argument is ignored" << std::endl;
+                }
+                ast_queries = unfoldtacpn::parse_string_queries(builder, qfile);
+            }
+            if (ast_queries.empty()) {
+                std::cerr << "There was an error parsing " << queryFile << std::endl;
+                std::exit(-1);
+            }
+
+            if(!options.getOutputQueryFile().empty())
+            {
+                std::fstream of(options.getOutputQueryFile(), std::ios::out);
+                unfoldtacpn::PQL::to_xml(of, ast_queries);
+            }
+            return std::unique_ptr<Query>(AST::toAST(ast_queries[0].first));
+        }
+
+    } catch (...) {
+        std::cout << "There was an error parsing the query file." << std::endl;
+        std::exit(-1);
+    }
+    return nullptr;
+}
 
 int main(int argc, char *argv[]) {
     srand(time(nullptr));
 
     ArgsParser parser;
     VerificationOptions options = parser.parse(argc, argv);
-    TAPNModelBuilder modelBuilder;
-    TAPN::TimedArcPetriNet *tapn;
-    std::string queryFile;
 
-    if(!options.getOutputModelFile().empty() && !options.getOutputQueryFile().empty()){
+    std::unique_ptr<TAPN::TimedArcPetriNet> tapn;
 
-        unfoldtacpn::ColoredPetriNetBuilder builder;
+    unfoldtacpn::ColoredPetriNetBuilder builder;
+    std::vector<int> initialPlacement;
+    {
+        TAPNModelBuilder modelBuilder;
         std::ifstream mf(options.getInputFile());
         builder.parseNet(mf);
         builder.unfold(modelBuilder);
         mf.close();
+        tapn.reset(modelBuilder.make_tapn());
+        initialPlacement = modelBuilder.initialMarking();
     }
 
     tapn->initialize(options.getGlobalMaxConstantsEnabled(), options.getGCDLowerGuardsEnabled());
@@ -41,42 +88,35 @@ int main(int argc, char *argv[]) {
         std::cout << std::endl;
         return 0;
     }
-    std::vector<int> initialPlacement = modelBuilder.initialMarking();
-    AST::Query *query = nullptr;
+
+
+    std::unique_ptr<AST::Query> query = nullptr;
     if (options.getWorkflowMode() == VerificationOptions::WORKFLOW_SOUNDNESS ||
         options.getWorkflowMode() == VerificationOptions::WORKFLOW_STRONG_SOUNDNESS) {
         if (options.getGCDLowerGuardsEnabled()) {
             std::cout << "Workflow-analysis does not support GCD-lowering" << std::endl;
-           std::exit(1);
+            std::exit(1);
         }
 
         if (options.getSearchType() != VerificationOptions::DEFAULT) {
             std::cout << "Workflow-analysis only supports the default search-strategy" << std::endl;
-           std::exit(1);
+            std::exit(1);
         }
 
         if (options.getQueryFile() != "") {
             std::cout << "Workflow-analysis does not accept a query file" << std::endl;
-           std::exit(1);
+            std::exit(1);
         }
         if (options.getWorkflowMode() == VerificationOptions::WORKFLOW_SOUNDNESS) {
             options.setSearchType(VerificationOptions::MINDELAYFIRST);
-            query = new AST::Query(AST::EF, new AST::BoolExpression(true));
+            query = std::make_unique<AST::Query>(AST::EF, new AST::BoolExpression(true));
         } else if (options.getWorkflowMode() == VerificationOptions::WORKFLOW_STRONG_SOUNDNESS) {
             options.setSearchType(VerificationOptions::DEPTHFIRST);
-            query = new AST::Query(AST::AF, new AST::BoolExpression(false));
+            query = std::make_unique<AST::Query>(AST::AF, new AST::BoolExpression(false));
         }
 
     } else {
-        try {
-            TAPNQueryParser queryParser(*tapn);
-            queryParser.parse(queryFile);
-            query = queryParser.getAST();
-
-        } catch (...) {
-            std::cout << "There was an error parsing the query file." << std::endl;
-            return 1;
-        }
+        query = parse_queries(options, builder);
 
         if (options.getTrace() != VerificationOptions::NO_TRACE &&
             (query->getQuantifier() == AST::CF || query->getQuantifier() == AST::CG)) {
@@ -86,11 +126,11 @@ int main(int argc, char *argv[]) {
 
         if (options.getTrace() == VerificationOptions::FASTEST_TRACE &&
             (options.getSearchType() != VerificationOptions::DEFAULT ||
-             query->getQuantifier() == AST::EG || query->getQuantifier() == AST::AF ||
-             options.getVerificationType() == VerificationOptions::TIMEDART)) {
+            query->getQuantifier() == AST::EG || query->getQuantifier() == AST::AF ||
+            options.getVerificationType() == VerificationOptions::TIMEDART)) {
             std::cout
-                    << "Fastest trace-option is only supported for reachability queries with default search strategy and without time darts."
-                    << std::endl;
+                << "Fastest trace-option is only supported for reachability queries with default search strategy and without time darts."
+                << std::endl;
             return 1;
         } else if (options.getTrace() == VerificationOptions::FASTEST_TRACE) {
             options.setSearchType(VerificationOptions::MINDELAYFIRST);
@@ -108,13 +148,10 @@ int main(int argc, char *argv[]) {
         options.setKeepDeadTokens(true);
     }
 
-    tapn->updatePlaceTypes(query, options);
+    tapn->updatePlaceTypes(query.get(), options);
 
-    int result = DiscreteVerification::DiscreteVerification::run(*tapn, initialPlacement, query, options);
+    int result = DiscreteVerification::DiscreteVerification::run(*tapn, initialPlacement, query.get(), options);
 
-    // cleanup
-    delete tapn;
-    delete query;
     return result;
 }
